requirements.txt
yfinance
pandas
numpy
requests
matplotlib
mplfinance
python-dotenv